requirements.txt
mysql-connector-python==8.2.0
python-dotenv==1.0.0
requests==2.31.0
flask==3.0.0
flask-cors==4.0.0
binance-connector==3.4.0
apscheduler==3.10.4
pandas==2.1.0
numpy==1.24.3
scipy==1.11.0
statsmodels==0.14.0
scikit-learn==1.3.0
yfinance==0.2.28
gunicorn==21.2.0
